requirements.txt
amqp==2.4.2
billiard==3.6.0.0

certifi==2019.3.9
chardet==3.0.4
Click==7.0
environs==7.1.0
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.5.0
MarkupSafe==1.1.1
marshmallow==3.3.0
oauthlib==3.0.1
pymongo==3.8.0
python-dotenv==0.10.3
pytz==2019.1
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.2.0
urllib3==1.25.2
vine==1.3.0
Werkzeug==0.15.4
dnspython==2.1.0
common==0.1.2
